JS init: fix brush loading to allow multiple uses of the same brush
The previous code immediately called the callback when another load of
the brush was already in progress, even if the load was not complete
yet. Replace with logic based on jQuery.Deferred that avoids this kind
of problem -- all callbacks are called only when the load has finished.

Ticket: none
Release-Note(fix,poweruser) {
  Correct loading function of syntax highlighter to support multiple
  uses of the same language/highlighter in a single page.
}